# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "HospitalShield AI API",
  "description": "Climate-resilient hospital digital twin (HC-03 planning prototype, synthetic data).",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/simulations": {
   "post": {
    "summary": "Create Simulation",
    "operationId": "create_simulation_api_v1_simulations_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SimulationRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SimulationResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/comparisons": {
   "post": {
    "summary": "Create Comparison",
    "operationId": "create_comparison_api_v1_comparisons_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ComparisonRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ComparisonResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/plans": {
   "post": {
    "summary": "Create Plan",
    "operationId": "create_plan_api_v1_plans_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PlanRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PlanResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
   "AccessIn": {
    "properties": {
     "start_hour": {
      "type": "integer",
      "maximum": 71.0,
      "minimum": 0.0,
      "title": "Start Hour",
      "default": 18
     },
     "duration_hours": {
      "type": "integer",
      "maximum": 72.0,
      "minimum": 0.0,
      "title": "Duration Hours",
      "default": 24
     },
     "staff_access_fraction": {
      "type": "number",
      "maximum": 1.0,
      "minimum": 0.0,
      "title": "Staff Access Fraction",
      "default": 0.6
     },
     "delivery_fraction": {
      "type": "number",
      "maximum": 1.0,
      "minimum": 0.0,
      "title": "Delivery Fraction",
      "default": 0.25
     },
     "arrival_multiplier": {
      "type": "number",
      "maximum": 3.0,
      "minimum": 0.0,
      "title": "Arrival Multiplier",
      "default": 1.4
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "AccessIn"
   },
...
   "AlertEpisode": {
    "properties": {
     "rule_id": {
      "type": "string",
      "title": "Rule Id"
     },
     "rule_text": {
      "type": "string",
      "title": "Rule Text"
     },
     "severity": {
      "type": "string",
      "title": "Severity"
     },
     "metric": {
      "type": "string",
      "title": "Metric"
     },
     "unit": {
      "type": "string",
      "title": "Unit"
     },
     "observed_peak": {
      "type": "number",
      "title": "Observed Peak"
     },
     "threshold": {
      "type": "number",
      "title": "Threshold"
     },
     "first_hour": {
      "type": "integer",
      "title": "First Hour"
     },
     "last_hour": {
      "type": "integer",
      "title": "Last Hour"
     },
     "duration_hours": {
      "type": "integer",
      "title": "Duration Hours"
     },
     "affected_variable": {
      "type": "string",
      "title": "Affected Variable"
     }
    },
    "additionalProperties": true,
    "type": "object",
    "required": [
     "rule_id",
     "rule_text",
     "severity",
     "metric",
     "unit",
     "observed_peak",
     "threshold",
     "first_hour",
     "last_hour",
     "duration_hours",
     "affected_variable"
    ],
    "title": "AlertEpisode"
   },
   "AttributionEntry": {
    "properties": {
     "intervention": {
      "type": "string",
      "title": "Intervention"
     },
     "detail": {
      "type": "object",
      "title": "Detail"
     },
     "deltas": {
      "additionalProperties": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/MetricDelta"
        },
        {
         "additionalProperties": {
          "$ref": "#/components/schemas/MetricDelta"
         },
         "type": "object"
        }
       ]
      },
      "type": "object",
      "title": "Deltas"
     },
     "explanation": {
      "type": "string",
      "title": "Explanation"
     }
    },
    "additionalProperties": true,
    "type": "object",
    "required": [
     "intervention",
     "detail",
     "deltas",
     "explanation"
    ],
    "title": "AttributionEntry"
   },
   "BackupIn": {
    "properties": {
     "generator_availability_fraction": {
      "type": "number",
      "maximum": 1.0,
      "minimum": 0.0,
      "title": "Generator Availability Fraction",
      "default": 1.0
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "BackupIn"
   },
   "ComparisonRequest": {
    "properties": {
     "scenario": {
      "type": "string",
      "enum": [
       "normal_operations",
       "heatwave_power_stress",
       "heatwave_power_outage",
       "flood_storm_access"
      ],
      "title": "Scenario"
     },
     "duration_hours": {
      "type": "integer",
      "maximum": 72.0,
      "minimum": 1.0,
      "title": "Duration Hours",
      "default": 72
     },
     "weather": {
      "$ref": "#/components/schemas/WeatherIn"
     },
     "stress": {
      "$ref": "#/components/schemas/StressIn"
     },
     "outage": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/OutageIn"
       },
       {
        "type": "null"
       }
      ]
     },
     "access": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/AccessIn"
       },
       {
        "type": "null"
       }
      ]
     },
     "backup": {
      "$ref": "#/components/schemas/BackupIn"
     },
     "interventions": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/InterventionsIn"
       },
       {
        "type": "null"
       }
      ]
     },
     "stochastic": {
      "type": "boolean",
      "title": "Stochastic",
      "default": false
     },
     "seed": {
      "type": "integer",
      "maximum": 2147483647.0,
      "minimum": 0.0,
      "title": "Seed",
      "default": 42
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "scenario"
    ],
    "title": "ComparisonRequest"
   },
   "ComparisonResponse": {
    "properties": {
     "model_version": {
      "type": "string",
      "title": "Model Version"
     },
     "input_hash": {
      "type": "string",
      "title": "Input Hash"
     },
     "seed": {
      "type": "integer",
      "title": "Seed"
     },
     "baseline": {
      "$ref": "#/components/schemas/SimulationResponse"
     },
     "treatment": {
      "$ref": "#/components/schemas/SimulationResponse"
     },
     "deltas": {
      "type": "object",
      "title": "Deltas"
     },
     "key_metric_deltas": {
      "type": "object",
      "title": "Key Metric Deltas"
     },
     "attribution": {
      "items": {
       "$ref": "#/components/schemas/AttributionEntry"
      },
      "type": "array",
      "title": "Attribution"
     },
     "explanations": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Explanations"
     }
    },
    "additionalProperties": true,
    "type": "object",
    "required": [
     "model_version",
     "input_hash",
     "seed",
     "baseline",
     "treatment",
     "deltas",
     "key_metric_deltas",
     "attribution",
     "explanations"
    ],
    "title": "ComparisonResponse"
   },
   "FailurePoint": {
    "properties": {
     "rank": {
      "type": "integer",
      "title": "Rank"
     },
     "resource": {
      "type": "string",
      "title": "Resource"
     },
     "rule_id": {
      "type": "string",
      "title": "Rule Id"
     },
     "first_hour": {
      "type": "integer",
      "title": "First Hour"
     },
     "last_hour": {
      "type": "integer",
      "title": "Last Hour"
     },
     "peak_severity": {
      "type": "string",
      "title": "Peak Severity"
     },
     "margin_trace": {
      "items": {
       "$ref": "#/components/schemas/MarginPoint"
      },
      "type": "array",
      "title": "Margin Trace"
     },
     "explanation": {
      "type": "string",
      "title": "Explanation"
     },
     "cascade_parent": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cascade Parent"
     },
     "cascade_children": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Cascade Children",
      "default": []
     }
    },
    "additionalProperties": true,
    "type": "object",
    "required": [
     "rank",
     "resource",
     "rule_id",
     "first_hour",
     "last_hour",
     "peak_severity",
     "margin_trace",
     "explanation"
    ],
    "title": "FailurePoint"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "HourRow": {
    "properties": {
     "hour": {
      "type": "integer",
      "title": "Hour"
     },
     "temperature_c": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Temperature C"
     },
     "grid_fraction": {
      "type": "number",
      "title": "Grid Fraction"
     },
     "power": {
      "$ref": "#/components/schemas/PowerRow"
     },
     "staffing": {
      "additionalProperties": {
       "$ref": "#/components/schemas/StaffingRow"
      },
      "type": "object",
      "title": "Staffing"
     },
     "units": {
      "additionalProperties": {
       "$ref": "#/components/schemas/UnitRow"
      },
      "type": "object",
      "title": "Units"
     },
     "supplies": {
      "additionalProperties": {
       "$ref": "#/components/schemas/SupplyRow"
      },
      "type": "object",
      "title": "Supplies",
      "default": {}
     },
     "access": {
      "$ref": "#/components/schemas/AccessRow",
      "default": {
       "staff_access_fraction": 1.0,
       "delivery_fraction": 1.0,
       "arrival_multiplier": 1.0
      }
     },
     "ed": {
      "additionalProperties": {
       "type": "integer"
      },
      "type": "object",
      "title": "Ed"
     },
     "discharged": {
      "type": "integer",
      "title": "Discharged",
      "default": 0
     },
     "binding_constraint": {
      "type": "string",
      "title": "Binding Constraint"
     },
     "margins": {
      "anyOf": [
       {
        "additionalProperties": {
         "type": "number"
        },
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Margins"
     }
    },
    "additionalProperties": true,
    "type": "object",
    "required": [
     "hour",
     "grid_fraction",
     "power",
     "staffing",
     "units",
     "ed",
     "binding_constraint"
    ],
    "title": "HourRow"
   },
   "InterventionsIn": {
    "properties": {
     "activate_surge_beds": {
      "items": {
       "$ref": "#/components/schemas/SurgeIn"
      },
      "type": "array",
      "title": "Activate Surge Beds"
     },
     "reallocate_staff": {
      "items": {
       "$ref": "#/components/schemas/StaffMoveIn"
      },
      "type": "array",
      "title": "Reallocate Staff"
     },
     "reduce_noncritical_load_kw": {
      "items": {
       "$ref": "#/components/schemas/ShedIn"
      },
      "type": "array",
      "title": "Reduce Noncritical Load Kw"
     },
     "recall_staff": {
      "items": {
       "$ref": "#/components/schemas/RecallIn"
      },
      "type": "array",
      "title": "Recall Staff"
     },
     "emergency_resupply": {
      "items": {
       "$ref": "#/components/schemas/ResupplyIn"
      },
      "type": "array",
      "title": "Emergency Resupply"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "InterventionsIn"
   },
...
   "OutageIn": {
    "properties": {
     "start_hour": {
      "type": "integer",
      "maximum": 71.0,
      "minimum": 0.0,
      "title": "Start Hour",
      "default": 24
     },
     "duration_hours": {
      "type": "integer",
      "maximum": 72.0,
      "minimum": 0.0,
      "title": "Duration Hours",
      "default": 36
     },
     "grid_fraction": {
      "type": "number",
      "maximum": 1.0,
      "minimum": 0.0,
      "title": "Grid Fraction",
      "default": 0.0
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "OutageIn"
   },
   "PlanRequest": {
    "properties": {
     "scenario": {
      "type": "string",
      "enum": [
       "normal_operations",
       "heatwave_power_stress",
       "heatwave_power_outage",
       "flood_storm_access"
      ],
      "title": "Scenario"
     },
     "duration_hours": {
      "type": "integer",
      "maximum": 72.0,
      "minimum": 1.0,
      "title": "Duration Hours",
      "default": 72
     },
     "weather": {
      "$ref": "#/components/schemas/WeatherIn"
     },
     "stress": {
      "$ref": "#/components/schemas/StressIn"
     },
     "outage": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/OutageIn"
       },
       {
        "type": "null"
       }
      ]
     },
     "access": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/AccessIn"
       },
       {
        "type": "null"
       }
      ]
     },
     "backup": {
      "$ref": "#/components/schemas/BackupIn"
     },
     "interventions": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/InterventionsIn"
       },
       {
        "type": "null"
       }
      ]
     },
     "stochastic": {
      "type": "boolean",
      "title": "Stochastic",
      "default": false
     },
     "seed": {
      "type": "integer",
      "maximum": 2147483647.0,
      "minimum": 0.0,
      "title": "Seed",
      "default": 42
     },
     "objective": {
      "$ref": "#/components/schemas/PlannerObjectiveIn"
     },
     "budget": {
      "$ref": "#/components/schemas/PlannerBudgetIn"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "scenario"
    ],
    "title": "PlanRequest"
   },
   "PlanResponse": {
    "properties": {
     "model_version": {
      "type": "string",
      "title": "Model Version"
     },
     "input_hash": {
      "type": "string",
      "title": "Input Hash"
     },
     "seed": {
      "type": "integer",
      "title": "Seed"
     },
     "label": {
      "type": "string",
      "title": "Label"
     },
     "action_plan": {
      "items": {
       "$ref": "#/components/schemas/PlanStep"
      },
      "type": "array",
      "title": "Action Plan"
     },
     "score": {
      "type": "object",
      "title": "Score"
     },
     "references": {
      "items": {
       "$ref": "#/components/schemas/ReferencePlan"
      },
      "type": "array",
      "title": "References"
     },
     "search_stats": {
      "type": "object",
      "title": "Search Stats"
     },
     "failure_points_before": {
      "items": {
       "$ref": "#/components/schemas/FailurePoint"
      },
      "type": "array",
      "title": "Failure Points Before"
     },
     "failure_points_after": {
      "items": {
       "$ref": "#/components/schemas/FailurePoint"
      },
      "type": "array",
      "title": "Failure Points After"
     },
     "assumptions": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Assumptions"
     },
     "provenance": {
      "$ref": "#/components/schemas/Provenance"
     },
     "limitations": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Limitations"
     }
    },
    "additionalProperties": true,
    "type": "object",
    "required": [
     "model_version",
     "input_hash",
     "seed",
     "label",
     "action_plan",
     "score",
     "references",
     "search_stats",
     "failure_points_before",
     "failure_points_after",
     "assumptions",
     "provenance",
     "limitations"
    ],
    "title": "PlanResponse"
   },
...
   "Provenance": {
    "properties": {
     "hospital_data": {
      "type": "string",
      "title": "Hospital Data",
      "default": "SYNTHETIC"
     },
     "weather": {
      "type": "string",
      "title": "Weather",
      "default": "SYNTHETIC_PROFILE"
     },
     "outputs": {
      "type": "string",
      "title": "Outputs",
      "default": "CALCULATED"
     },
     "model_version": {
      "type": "string",
      "title": "Model Version"
     },
     "warnings": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Warnings",
      "default": []
     }
    },
    "additionalProperties": true,
    "type": "object",
    "required": [
     "model_version"
    ],
    "title": "Provenance"
   },
...
   "SimulationRequest": {
    "properties": {
     "scenario": {
      "type": "string",
      "enum": [
       "normal_operations",
       "heatwave_power_stress",
       "heatwave_power_outage",
       "flood_storm_access"
      ],
      "title": "Scenario"
     },
     "duration_hours": {
      "type": "integer",
      "maximum": 72.0,
      "minimum": 1.0,
      "title": "Duration Hours",
      "default": 72
     },
     "weather": {
      "$ref": "#/components/schemas/WeatherIn"
     },
     "stress": {
      "$ref": "#/components/schemas/StressIn"
     },
     "outage": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/OutageIn"
       },
       {
        "type": "null"
       }
      ]
     },
     "access": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/AccessIn"
       },
       {
        "type": "null"
       }
      ]
     },
     "backup": {
      "$ref": "#/components/schemas/BackupIn"
     },
     "interventions": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/InterventionsIn"
       },
       {
        "type": "null"
       }
      ]
     },
     "stochastic": {
      "type": "boolean",
      "title": "Stochastic",
      "default": false
     },
     "seed": {
      "type": "integer",
      "maximum": 2147483647.0,
      "minimum": 0.0,
      "title": "Seed",
      "default": 42
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "scenario"
    ],
    "title": "SimulationRequest"
   },
   "SimulationResponse": {
    "properties": {
     "run_id": {
      "type": "string",
      "title": "Run Id"
     },
     "model_version": {
      "type": "string",
      "title": "Model Version"
     },
     "input_hash": {
      "type": "string",
      "title": "Input Hash"
     },
     "seed": {
      "type": "integer",
      "title": "Seed"
     },
     "stochastic": {
      "type": "boolean",
      "title": "Stochastic"
     },
     "summary": {
      "$ref": "#/components/schemas/Summary"
     },
     "series": {
      "items": {
       "$ref": "#/components/schemas/HourRow"
      },
      "type": "array",
      "title": "Series"
     },
     "alerts": {
      "items": {
       "$ref": "#/components/schemas/AlertEpisode"
      },
      "type": "array",
      "title": "Alerts"
     },
     "failure_points": {
      "items": {
       "$ref": "#/components/schemas/FailurePoint"
      },
      "type": "array",
      "title": "Failure Points"
     },
     "assumptions": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Assumptions"
     },
     "provenance": {
      "$ref": "#/components/schemas/Provenance"
     },
     "limitations": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Limitations"
     }
    },
    "additionalProperties": true,
    "type": "object",
    "required": [
     "run_id",
     "model_version",
     "input_hash",
     "seed",
     "stochastic",
     "summary",
     "series",
     "alerts",
     "failure_points",
     "assumptions",
     "provenance",
     "limitations"
    ],
    "title": "SimulationResponse"
   },
...
   "StressIn": {
    "properties": {
     "arrival_multiplier": {
      "type": "number",
      "maximum": 3.0,
      "minimum": 0.0,
      "title": "Arrival Multiplier",
      "default": 1.25
     },
     "los_multiplier": {
      "type": "number",
      "maximum": 2.0,
      "minimum": 0.5,
      "title": "Los Multiplier",
      "default": 1.0
     },
     "staff_availability_fraction": {
      "type": "number",
      "maximum": 1.0,
      "minimum": 0.0,
      "title": "Staff Availability Fraction",
      "default": 0.9
     },
     "power_supply_fraction": {
      "type": "number",
      "maximum": 1.0,
      "minimum": 0.0,
      "title": "Power Supply Fraction",
      "default": 0.8
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "StressIn"
   },
   "Summary": {
    "properties": {
     "peak_ward_occupancy_pct": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Peak Ward Occupancy Pct"
     },
     "peak_icu_occupancy_pct": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Peak Icu Occupancy Pct"
     },
     "peak_ed_occupancy_pct": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Peak Ed Occupancy Pct"
     },
     "hours_ward_above_threshold": {
      "type": "integer",
      "title": "Hours Ward Above Threshold"
     },
     "hours_icu_above_threshold": {
      "type": "integer",
      "title": "Hours Icu Above Threshold"
     },
     "peak_ed_waiting": {
      "type": "integer",
      "title": "Peak Ed Waiting"
     },
     "peak_ed_boarding": {
      "type": "integer",
      "title": "Peak Ed Boarding"
     },
     "overflow_patient_hours": {
      "type": "integer",
      "title": "Overflow Patient Hours"
     },
     "staff_shortfall_nurse_hours": {
      "type": "integer",
      "title": "Staff Shortfall Nurse Hours"
     },
     "hours_beds_over_usable": {
      "type": "integer",
      "title": "Hours Beds Over Usable"
     },
     "peak_power_demand_kw": {
      "type": "number",
      "title": "Peak Power Demand Kw"
     },
     "hours_power_deficit": {
      "type": "integer",
      "title": "Hours Power Deficit"
     },
     "peak_power_deficit_kw": {
      "type": "number",
      "title": "Peak Power Deficit Kw"
     },
     "energy_unserved_kwh": {
      "type": "number",
      "title": "Energy Unserved Kwh"
     },
     "min_power_reserve_kw": {
      "type": "number",
      "title": "Min Power Reserve Kw"
     },
     "hours_critical_load_at_risk": {
      "type": "integer",
      "title": "Hours Critical Load At Risk"
     },
     "fuel_remaining_l": {
      "type": "number",
      "title": "Fuel Remaining L"
     },
     "hours_to_fuel_exhaustion": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Hours To Fuel Exhaustion"
     },
     "resilience_index": {
      "type": "number",
      "title": "Resilience Index"
     },
     "first_failure": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/FirstFailure"
       },
       {
        "type": "null"
       }
      ]
     },
     "time_to_first_failure_hours": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Time To First Failure Hours"
     },
     "root_failures": {
      "type": "integer",
      "title": "Root Failures",
      "default": 0
     },
     "cascade_children": {
      "type": "integer",
      "title": "Cascade Children",
      "default": 0
     }
    },
    "additionalProperties": true,
    "type": "object",
    "required": [
     "hours_ward_above_threshold",
     "hours_icu_above_threshold",
     "peak_ed_waiting",
     "peak_ed_boarding",
     "overflow_patient_hours",
     "staff_shortfall_nurse_hours",
     "hours_beds_over_usable",
     "peak_power_demand_kw",
     "hours_power_deficit",
     "peak_power_deficit_kw",
     "energy_unserved_kwh",
     "min_power_reserve_kw",
     "hours_critical_load_at_risk",
     "fuel_remaining_l",
     "resilience_index"
    ],
    "title": "Summary"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "WeatherIn": {
    "properties": {
     "mode": {
      "type": "string",
      "const": "synthetic_profile",
      "title": "Mode",
      "default": "synthetic_profile"
     },
     "max_temperature_c": {
      "type": "number",
      "maximum": 55.0,
      "minimum": -10.0,
      "title": "Max Temperature C",
      "default": 41
     },
     "min_temperature_c": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 55.0,
        "minimum": -10.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Min Temperature C"
     },
     "humidity_pct": {
      "type": "number",
      "maximum": 100.0,
      "minimum": 0.0,
      "title": "Humidity Pct",
      "default": 55
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "WeatherIn"
   }
  }
 }
}